+++ main.tf
@@ -44,7 +44,7 @@ resource "azurerm_kubernetes_cluster" "main" {
     avm_git_last_modified_at = "2023-11-13 03:03:50"
     avm_git_org              = "Azure"
     avm_git_repo             = "terraform-azurerm-aks"
-    avm_yor_trace            = "ad9ad71d-4be6-4dbb-b5ca-b43d707fdd5c"
+    avm_yor_trace            = "924f2b15-7493-4e0f-a761-192853d29027"
     } /*<box>*/ : replace(k, "avm_", var.tracing_tags_prefix) => v } : {}) /*</box>*/), (/*<box>*/ (var.tracing_tags_enabled ? { for k, v in /*</box>*/ {
     avm_yor_name = "main"
   } /*<box>*/ : replace(k, "avm_", var.tracing_tags_prefix) => v } : {}) /*</box>*/))
@@ -330,13 +330,13 @@ resource "azurerm_kubernetes_cluster" "main" {
     }
   }
   dynamic "ingress_application_gateway" {
-    for_each = var.ingress_application_gateway_enabled ? ["ingress_application_gateway"] : []
+    for_each = local.ingress_application_gateway_enabled ? ["ingress_application_gateway"] : []
 
     content {
-      gateway_id   = var.ingress_application_gateway_id
-      gateway_name = var.ingress_application_gateway_name
-      subnet_cidr  = var.ingress_application_gateway_subnet_cidr
-      subnet_id    = var.ingress_application_gateway_subnet_id
+      gateway_id   = try(var.brown_field_application_gateway_for_ingress.id, null)
+      gateway_name = try(var.green_field_application_gateway_for_ingress.name, null)
+      subnet_cidr  = try(var.green_field_application_gateway_for_ingress.subnet_cidr, null)
+      subnet_id    = try(var.green_field_application_gateway_for_ingress.subnet_id, null)
     }
   }
   dynamic "key_management_service" {
@@ -575,6 +575,10 @@ resource "azurerm_kubernetes_cluster" "main" {
       condition     = !var.enable_auto_scaling || var.agents_type == "VirtualMachineScaleSets"
       error_message = "Autoscaling on default node pools is only supported when the Kubernetes Cluster is using Virtual Machine Scale Sets type nodes."
     }
+    precondition {
+      condition     = var.brown_field_application_gateway_for_ingress == null || var.green_field_application_gateway_for_ingress == null
+      error_message = "Either one of `var.brown_field_application_gateway_for_ingress` or `var.green_field_application_gateway_for_ingress` must be `null`."
+    }
   }
 }
 
@@ -640,7 +644,7 @@ resource "azurerm_kubernetes_cluster_node_pool" "node_pool" {
     avm_git_last_modified_at = "2023-05-04 05:02:32"
     avm_git_org              = "Azure"
     avm_git_repo             = "terraform-azurerm-aks"
-    avm_yor_trace            = "5fb71500-0879-4743-b571-b73b4d29f6f5"
+    avm_yor_trace            = "2409ad32-ea11-4e74-b160-a5357c5f19b5"
     } /*<box>*/ : replace(k, "avm_", var.tracing_tags_prefix) => v } : {}) /*</box>*/), (/*<box>*/ (var.tracing_tags_enabled ? { for k, v in /*</box>*/ {
     avm_yor_name = "node_pool"
   } /*<box>*/ : replace(k, "avm_", var.tracing_tags_prefix) => v } : {}) /*</box>*/))
@@ -784,7 +788,7 @@ resource "azurerm_log_analytics_workspace" "main" {
     avm_git_last_modified_at = "2023-06-05 02:21:33"
     avm_git_org              = "Azure"
     avm_git_repo             = "terraform-azurerm-aks"
-    avm_yor_trace            = "9ccc1f0e-173a-4d2d-a9c8-12f4f06d073b"
+    avm_yor_trace            = "8393c5bc-eb70-4793-88b1-98fb2149983e"
     } /*<box>*/ : replace(k, "avm_", var.tracing_tags_prefix) => v } : {}) /*</box>*/), (/*<box>*/ (var.tracing_tags_enabled ? { for k, v in /*</box>*/ {
     avm_yor_name = "main"
   } /*<box>*/ : replace(k, "avm_", var.tracing_tags_prefix) => v } : {}) /*</box>*/))
@@ -825,7 +829,7 @@ resource "azurerm_log_analytics_solution" "main" {
     avm_git_last_modified_at = "2023-07-20 06:04:07"
     avm_git_org              = "Azure"
     avm_git_repo             = "terraform-azurerm-aks"
-    avm_yor_trace            = "69f8a9e4-e374-4524-98db-55e64f589cf1"
+    avm_yor_trace            = "b09b4cf8-372d-4c34-8e9c-0eea57305550"
     } /*<box>*/ : replace(k, "avm_", var.tracing_tags_prefix) => v } : {}) /*</box>*/), (/*<box>*/ (var.tracing_tags_enabled ? { for k, v in /*</box>*/ {
     avm_yor_name = "main"
   } /*<box>*/ : replace(k, "avm_", var.tracing_tags_prefix) => v } : {}) /*</box>*/))
@@ -887,4 +891,78 @@ resource "azurerm_role_assignment" "network_contributor_on_subnet" {
       error_message = "Cannot set both of `var.create_role_assignment_network_contributor` and `var.network_contributor_role_assigned_subnet_ids`."
     }
   }
+}
+
+data "azurerm_client_config" "this" {}
+
+data "azurerm_virtual_network" "application_gateway_vnet" {
+  count = var.create_role_assignments_for_application_gateway && local.use_brown_field_gw_for_ingress ? 1 : 0
+
+  name                = local.existing_application_gateway_subnet_vnet_name
+  resource_group_name = local.existing_application_gateway_subnet_resource_group_name
+}
+
+resource "azurerm_role_assignment" "application_gateway_existing_vnet_network_contributor" {
+  count = var.create_role_assignments_for_application_gateway && local.use_brown_field_gw_for_ingress ? 1 : 0
+
+  principal_id         = azurerm_kubernetes_cluster.main.ingress_application_gateway[0].ingress_application_gateway_identity[0].object_id
+  scope                = data.azurerm_virtual_network.application_gateway_vnet[0].id
+  role_definition_name = "Network Contributor"
+
+  lifecycle {
+    precondition {
+      condition     = data.azurerm_client_config.this.subscription_id == local.existing_application_gateway_subnet_subscription_id_for_ingress
+      error_message = "Application Gateway's subnet must be in the same subscription, or `var.application_gateway_for_ingress.create_role_assignments` must be set to `false`."
+    }
+  }
+}
+
+resource "azurerm_role_assignment" "application_gateway_byo_vnet_network_contributor" {
+  count = var.create_role_assignments_for_application_gateway && local.use_green_field_gw_for_ingress ? 1 : 0
+
+  principal_id         = azurerm_kubernetes_cluster.main.ingress_application_gateway[0].ingress_application_gateway_identity[0].object_id
+  scope                = join("/", slice(local.default_nodepool_subnet_segments, 0, length(local.default_nodepool_subnet_segments) - 2))
+  role_definition_name = "Network Contributor"
+
+  lifecycle {
+    precondition {
+      condition     = var.green_field_application_gateway_for_ingress == null || !(var.create_role_assignments_for_application_gateway && var.vnet_subnet_id == null)
+      error_message = "When `var.vnet_subnet_id` is `null`, you must set `var.create_role_assignments_for_application_gateway` to `false`, set `var.green_field_application_gateway_for_ingress` to `null`."
+    }
+  }
+}
+
+resource "azurerm_role_assignment" "existing_application_gateway_contributor" {
+  count = var.create_role_assignments_for_application_gateway && local.use_brown_field_gw_for_ingress ? 1 : 0
+
+  principal_id         = azurerm_kubernetes_cluster.main.ingress_application_gateway[0].ingress_application_gateway_identity[0].object_id
+  scope                = var.brown_field_application_gateway_for_ingress.id
+  role_definition_name = "Contributor"
+
+  lifecycle {
+    precondition {
+      condition     = var.brown_field_application_gateway_for_ingress == null ? true : data.azurerm_client_config.this.subscription_id == local.existing_application_gateway_subscription_id_for_ingress
+      error_message = "Application Gateway must be in the same subscription, or `var.create_role_assignments_for_application_gateway` must be set to `false`."
+    }
+  }
+}
+
+data "azurerm_resource_group" "ingress_gw" {
+  count = var.create_role_assignments_for_application_gateway && local.use_brown_field_gw_for_ingress ? 1 : 0
+
+  name = local.existing_application_gateway_resource_group_for_ingress
+}
+
+data "azurerm_resource_group" "aks_rg" {
+  count = var.create_role_assignments_for_application_gateway ? 1 : 0
+
+  name = var.resource_group_name
+}
+
+resource "azurerm_role_assignment" "application_gateway_resource_group_reader" {
+  count = var.create_role_assignments_for_application_gateway && local.ingress_application_gateway_enabled ? 1 : 0
+
+  principal_id         = azurerm_kubernetes_cluster.main.ingress_application_gateway[0].ingress_application_gateway_identity[0].object_id
+  scope                = local.use_brown_field_gw_for_ingress ? data.azurerm_resource_group.ingress_gw[0].id : data.azurerm_resource_group.aks_rg[0].id
+  role_definition_name = "Reader"
 }